left-hand-image-with-caption version 1.0
Image with caption template for NB mailout, using tested Mailchimp code. Image takes 1/4 of the column and is aligned left; caption takes up 3/4 caption on right. Works in versions of Outlook after version 2003. Replace CHANGE-IMAGE-REFERENCE-HERE.jpg and CHANGE-CAPTION-CONTENT-HERE. The image should just be the filename of the file you upload to the NB newsletter's Files area.